% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ae}
\alias{ae}
\title{Absolute Error}
\usage{
ae(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector with ground truth values.}

\item{y_pred}{A numeric vector with predicted values.}
}
\description{
Absolute Error
}
