% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_assay}
\alias{aggregate_assay}
\title{Aggregate Seurat assay over groups}
\usage{
aggregate_assay(object, group_name, fun = "mean", assay = "RNA", slot = "data")
}
\arguments{
\item{group_name}{A character vector indicating the metadata column to aggregate over.}

\item{fun}{The summary function to be applied to each group.}

\item{assay}{The assay to summarize.}

\item{slot}{The slot to summarize.}
}
\value{
A Seurat object.
}
\description{
Aggregate Seurat assay over groups
}
