% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{coef.GRNData}
\alias{coef.GRNData}
\alias{coef.RegulatoryNetwork}
\alias{coef.Network}
\title{Get fitted coefficients}
\usage{
\method{coef}{GRNData}(object, network = DefaultNetwork(object))

\method{coef}{RegulatoryNetwork}(object, network = DefaultNetwork(object))

\method{coef}{Network}(object)
}
\description{
Get fitted coefficients
}
