% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute metrics for model evaluation}
\usage{
compute_metrics(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector with ground truth values.}

\item{y_pred}{A numeric vector with predicted values.}
}
\value{
A \code{data.frame} with computed metrics
}
\description{
Compute metrics for model evaluation
}
