% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{create_GRNData}
\alias{create_GRNData}
\title{Create GRNData object from a Seurat and a GRN object}
\usage{
create_GRNData(seurat_object, grn_object)
}
\description{
Create GRNData object from a Seurat and a GRN object
}
