% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_folds}
\alias{cv_folds}
\title{Create folds for cross validation}
\usage{
cv_folds(data, folds = 5, strata = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing the variables in the model.}

\item{folds}{Number of cross-validation folds.}

\item{strata}{A character vector with strata for stratified CV.}
}
\value{
A list with fold indices
}
\description{
Create folds for cross validation
}
