% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_model}
\alias{cv_model}
\title{Cross-validation for GRN inference models}
\usage{
cv_model(
  formula,
  data,
  method = c("glm", "glmnet", "cv.glmnet", "xgb", "bagging_ridge", "bayesian_ridge"),
  k_folds = 5,
  strata = NULL,
  ...
)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{method}{A character string indicating the method to fit the model.
\itemize{
\item \code{'glm'} - Generalized Liner Model with \code{\link[glm]{stats}}.
\item \code{'glmnet'}, \code{'cv.glmnet'} - Regularized Generalized Liner Model with \code{\link[glmnet]{glmnet}}.
\item \code{'xgb'} - Gradient Boosting Regression using \code{\link[xgboost]{xgboost}}.
\item \code{'bagging_ridge'} - Bagging Ridge Regression using scikit-learn via \link[xgboost]{reticulate}.
\item \code{'bayesian_ridge'} - Bayesian Ridge Regression using scikit-learn via \link[xgboost]{reticulate}.
}}

\item{k_folds}{Number of cross-validation folds.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Cross-validation for GRN inference models
}
