% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{DefaultNetwork}
\alias{DefaultNetwork}
\alias{DefaultNetwork.GRNData}
\alias{DefaultNetwork.RegulatoryNetwork}
\title{Get active network}
\usage{
DefaultNetwork(object, ...)

\method{DefaultNetwork}{GRNData}(object)

\method{DefaultNetwork}{RegulatoryNetwork}(object)
}
\description{
Get active network
}
