% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dMcast}
\alias{dMcast}
\title{Copy of the dMcast function from the Matrix.utils package,
since this is off CRAN and does not seem to be maintained anymore}
\usage{
dMcast(
  data,
  formula,
  fun.aggregate = "sum",
  value.var = NULL,
  as.factors = FALSE,
  factor.nas = TRUE,
  drop.unused.levels = TRUE
)
}
\description{
Copy of the dMcast function from the Matrix.utils package,
since this is off CRAN and does not seem to be maintained anymore
}
