% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/grn.R
\name{find_modules}
\alias{find_modules}
\alias{find_modules.RegulatoryNetwork}
\alias{find_modules.SeuratPlus}
\title{Find TF modules in regulatory network}
\usage{
find_modules(object, ...)

\method{find_modules}{RegulatoryNetwork}(
  object,
  p_thresh = 0.05,
  rsq_thresh = 0.1,
  nvar_thresh = 10,
  min_genes_per_module = 5
)

\method{find_modules}{SeuratPlus}(
  object,
  network = "glm_network",
  p_thresh = 0.05,
  rsq_thresh = 0.1,
  nvar_thresh = 10,
  min_genes_per_module = 5
)
}
\arguments{
\item{p_thresh}{Float indicating the significance threshold on the adjusted p-value.}

\item{rsq_thresh}{Float indicating the \eqn{R^2} threshold on the adjusted p-value.}

\item{nvar_thresh}{Integer indicating the minimum number of genes in a module.}
}
\value{
A RegulatoryNetwork object.

A Network object.
}
\description{
Find TF modules in regulatory network
}
