% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_peaks_near_genes}
\alias{find_peaks_near_genes}
\title{Find peaks or regions near gene body or TSS}
\usage{
find_peaks_near_genes(
  peaks,
  genes,
  upstream = 1e+05,
  downstream = 0,
  sep = c("-", "-"),
  only_tss = FALSE
)
}
\arguments{
\item{peaks}{A \code{GRanges} object with peak regions.}

\item{upstream}{Integer defining the distance upstream of the gene to consider.}

\item{downstream}{Integer defining the distance downstream of the gene to consider.}

\item{sep}{Vector of separators to use for genomic string.
First element is used to separate chromosome and coordinates,
second separator is used to separate start and end coordinates.}

\item{only_tss}{Logical. Measure distance from the TSS (\code{TRUE}) or from the entire gene body (\code{FALSE}).}

\item{gene}{A \code{GRanges} object with gene coordinates.}

\item{verbose}{Logical. Display messages}
}
\value{
A sparse binary Matrix with gene/peak matches.
}
\description{
Find peaks or regions near gene body or TSS
}
