% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_bayesian_ridge}
\alias{fit_bayesian_ridge}
\title{Fit a bayesian ridge regression model as implemented in scikit-learn (python)}
\usage{
fit_bayesian_ridge(formula, data, p_method = c("t", "wilcox"), ...)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{p_method}{The test used to calculate p-values. Options are 't' for \code{t.test}, and 'wilcox' for \code{wilcox.test}}

\item{...}{Other parameters for the model fitting function.}

\item{alpha}{Positive float indicating the regularization strength.}

\item{solver}{Solver to use in the computational routines.
Options include ‘auto’, ‘svd’, ‘cholesky’, ‘lsqr’, ‘sparse_cg’, ‘sag’, ‘saga’.}

\item{bagging_number}{The number of ridge regression model in the bagging.}

\item{n_jobs}{The number of cores used to fit the model.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Fit a bayesian ridge regression model as implemented in scikit-learn (python)
}
