% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_brms}
\alias{fit_brms}
\title{Fit a Bayesian regression model with brms and Stan}
\usage{
fit_brms(
  formula,
  data,
  family = gaussian,
  prior = brms::prior(normal(0, 1)),
  ...
)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{family}{A description of the error distribution and link function to be used in the model.
See \code{\link[family]{stats}} for mode details.}

\item{prior}{The prior distribution of the coefficients.
See \code{\link[set_prior]{brms}} for mode details.
The default (\code{prior(normal(0,1))}) results in ridge regularization.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Fit a Bayesian regression model with brms and Stan
}
