% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_cvglmnet}
\alias{fit_cvglmnet}
\title{Cross-validation for regularized generalized linear models}
\usage{
fit_cvglmnet(formula, data, family = gaussian, alpha = 0.5, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{family}{A description of the error distribution and link function to be used in the model.
See \code{\link[family]{stats}} for mode details.}

\item{alpha}{The elasticnet mixing parameter. See \code{\link[glmnet]{glmnet}} for details.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Cross-validation for regularized generalized linear models
}
