% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit generalized linear model}
\usage{
fit_glm(formula, data, family = gaussian, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{family}{A description of the error distribution and link function to be used in the model.
See \code{\link[family]{stats}} for mode details.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Fit generalized linear model
}
