% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/grn.R
\name{fit_grn_models}
\alias{fit_grn_models}
\alias{fit_grn_models.SeuratPlus}
\title{Fit models for gene expression}
\usage{
fit_grn_models(object, ...)

\method{fit_grn_models}{SeuratPlus}(
  object,
  genes = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  parallel = FALSE,
  tf_cor = 0.1,
  peak_cor = 0,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("glm", "glmnet", "cv.glmnet", "brms", "xgb", "bagging_ridge",
    "bayesian_ridge"),
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{...}{Other parameters for the model fitting function.}

\item{genes}{A character vector with the target genes to consider for GRN inference.
Takes all VariableFeatures in the object per default.}

\item{peak_to_gene_method}{Character specifying the method to
link peak overlapping motif regions to nearby genes. One of 'Signac' or 'GREAT'.}

\item{upstream}{Integer defining the distance upstream of the gene to consider as potential regulatory region.}

\item{downstream}{Integer defining the distance downstream of the gene to consider as potential regulatory region.}

\item{extend}{Integer defining the distance from the upstream and downstream of the basal regulatory region.
Only used of \code{peak_to_gene_method = 'GREAT'}.}

\item{only_tss}{Logical. Measure distance from the TSS (\code{TRUE}) or from the entire gene body (\code{FALSE}).}

\item{peak_to_gene_domains}{\code{GenomicRanges} object with regulatory regions for each gene.
If provided, all other peaks-to-gene methods are overwritten and these regions are used instead.}

\item{parallel}{Logical. Whether to parellelize the computation with \code{\link[foreach]{foreach}}.}

\item{tf_cor}{Threshold for TF - target gene correlation.}

\item{peak_cor}{Threshold for binding peak - target gene correlation.}

\item{method}{A character string indicating the method to fit the model.
\itemize{
\item \code{'glm'} - Generalized Liner Model with \code{\link[glm]{stats}}.
\item \code{'glmnet'}, \code{'cv.glmnet'} - Regularized Generalized Liner Model with \code{\link[glmnet]{glmnet}}.
\item \code{'brms'} - Bayesian Regression Models using \code{\link[brms-package]{brms}}.
\item \code{'xgb'} - Gradient Boosting Regression using \code{\link[xgboost]{xgboost}}.
\item \code{'bagging_ridge'} - Bagging Ridge Regression using scikit-learn via \link[reticulate]{reticulate}.
\item \code{'bayesian_ridge'} - Bayesian Ridge Regression using scikit-learn via \link[reticulate]{reticulate}.
}}

\item{interaction_term}{The interaction term to use in the model between TF and binding site.
\itemize{
\item \code{'+'} for additive interaction.
\item \code{':'} for 'multiplicative' interaction.
\item \code{'*'} for crossing interaction, i.e. additive AND 'multiplicative'.
For more info, see \code{\link[formula]{stats}}
}}

\item{adjust_method}{Method for adjusting p-values.}

\item{scale}{Logical. Whether to z-transform the expression and accessibility matrices.}

\item{verbose}{Logical. Display messages}
}
\value{
A SeuratPlus object.
}
\description{
Fit models for gene expression
}
