% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_model}
\alias{fit_model}
\title{Fit (regularized) generalized linear model}
\usage{
fit_model(formula, data, method = "glm", family = gaussian, alpha = 1, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{method}{A character string indicating the method to fit the model.
Possible values are \code{'glm'}, \code{'glmnet'} and \code{'cv.glmnet'}}

\item{family}{A description of the error distribution and link function to be used in the model.
See \code{\link[family]{stats}} for mode details.}

\item{alpha}{The elasticnet mixing parameter. See \code{\link[glmnet]{glmnet}} for details.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Fit (regularized) generalized linear model
}
