% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{format_coefs}
\alias{format_coefs}
\title{Format network coefficients}
\usage{
format_coefs(coefs, term = ":", adjust_method = "fdr")
}
\arguments{
\item{coefs}{A data frame with coefficients}
}
\value{
A data frame.
}
\description{
Format network coefficients
}
