% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{get_network_graph}
\alias{get_network_graph}
\alias{get_network_graph.SeuratPlus}
\title{Compute network graph embedding using UMAP.}
\usage{
get_network_graph(object, ...)

\method{get_network_graph}{SeuratPlus}(
  object,
  network = "glm_network",
  graph_name = "module_graph",
  rna_assay = "RNA",
  rna_slot = "data",
  edge_method = c("weighted", "corr", "coef"),
  features = NULL,
  random_seed = 111,
  ...
)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments for \code{\link[umap]{uwot}}.}

\item{network}{Name of the network to use.}

\item{graph_name}{Name of the graph.}

\item{rna_assay}{Name of the RNA assay.}

\item{rna_slot}{Name of the RNA slot to use.}

\item{edge_method}{Method to compute edge weights:
\itemize{
\item \code{'weighted'} - Correlation weighted by GRN coefficient.
\item \code{'corr'} - Only correlation.
\item \code{'coef'} - Only GRN coefficient.
}}
}
\value{
A SeuratPlus object.
}
\description{
Compute network graph embedding using UMAP.
}
