% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{get_tf_network}
\alias{get_tf_network}
\alias{get_tf_network.SeuratPlus}
\title{Get sub-network centered around one TF.}
\usage{
get_tf_network(object, ...)

\method{get_tf_network}{SeuratPlus}(
  object,
  tf,
  network = DefaultNetwork(object),
  graph = "module_graph",
  features = NULL,
  order = 3,
  keep_all_edges = FALSE,
  verbose = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{object}{An object.}

\item{tf}{The transcription factor to center around.}

\item{network}{Name of the network to use.}

\item{graph}{Name of the graph.}

\item{features}{Features to use. If \code{NULL} uses all features in the graph.}

\item{order}{Integer indicating the maximal order of the graph.}

\item{keep_all_edges}{Logical, whether to maintain all edges to each leaf
or prune to the strongest overall connection.}

\item{verbose}{Logical. Whether to print messages.}

\item{parallel}{Logical. Whether to parallelize the computation with \code{\link[foreach]{foreach}}.}
}
\value{
A SeuratPlus object.
}
\description{
Get sub-network centered around one TF.
}
