% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{GetAssay}
\alias{GetAssay}
\alias{GetAssay.GRNData}
\title{Get Seurat assay from GRNData}
\usage{
GetAssay(object, ...)

\method{GetAssay}{GRNData}(object, assay = NULL)
}
\description{
Get Seurat assay from GRNData
}
