% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{GetAssaySummary}
\alias{GetAssaySummary}
\alias{GetAssaySummary.Seurat}
\title{Get summary of seurat assay}
\usage{
GetAssaySummary(object, ...)

\method{GetAssaySummary}{Seurat}(object, group_name, assay = NULL, verbose = TRUE)
}
\description{
Get summary of seurat assay
}
