% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetNetwork.SeuratPlus}
\alias{GetNetwork.SeuratPlus}
\alias{GetNetwork.RegulatoryNetwork}
\title{Get network}
\usage{
\method{GetNetwork}{SeuratPlus}(object, network = "glm_network")

\method{GetNetwork}{RegulatoryNetwork}(object, network = "glm_network")
}
\description{
Get network

Get network
}
