% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{GetNetworkData}
\alias{GetNetworkData}
\alias{GetNetworkData.SeuratPlus}
\title{Get network data}
\usage{
GetNetworkData(object, ...)

\method{GetNetworkData}{SeuratPlus}(object)
}
\description{
Get network data
}
