% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{GRNData-class}
\alias{GRNData-class}
\alias{GRNData}
\title{The GRNData class}
\description{
The GRNData object is an extended \code{Seurat} object
for the storage and analysis of Regulatory network data.
}
\section{Slots}{

\describe{
\item{\code{grn}}{A named list containing \code{RegulatoryNetwork} objects with inferred networks.}
}}

