% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/inference.R
\name{infer_grn}
\alias{infer_grn}
\alias{infer_grn.SeuratPlus}
\title{Infer a Gene Regulatory Network with \code{Pando}}
\usage{
infer_grn(object, ...)

\method{infer_grn}{SeuratPlus}(
  object,
  upstream = 1e+05,
  downstream = 0,
  only_tss = FALSE,
  parallel = FALSE,
  tf_cor = 0.1,
  peak_cor = 0,
  method = "glm",
  alpha = 0.5,
  family = gaussian,
  interaction_term = ":",
  adjust_method = "fdr",
  verbose = TRUE
)
}
\arguments{
\item{upstream}{Integer defining the distance upstream of the gene to consider as potential regulatory region.}

\item{downstream}{Integer defining the distance downstream of the gene to consider as potential regulatory region.}

\item{only_tss}{Logical. Measure distance from the TSS (\code{TRUE}) or from the entire gene body (\code{FALSE}).}

\item{parallel}{Logical. Whether to parellelize the computation with \code{\link[foreach]{foreach}}.}

\item{tf_cor}{Threshold for TF - target gene correlation.}

\item{peak_cor}{Threshold for binding peak - target gene correlation.}

\item{method}{A character string indicating the method to fit the model.
Possible values are \code{'glm'}, \code{'glmnet'} and \code{'cv.glmnet'}.}

\item{alpha}{The elasticnet mixing parameter. See \code{\link[glmnet]{glmnet}} for details.}

\item{family}{A description of the error distribution and link function to be used in the model.
See \code{\link[family]{stats}} for mode details.}

\item{interaction_term}{The interaction term to use in the model between TF and binding site.
* \code{'+'} for additive interaction.
* \code{':'} for 'multiplicative' interaction.
* \code{'\*'} for crossing interaction, i.e. additive AND 'multiplicative'.
For more info, see \code{\link[formula]{stats}}}

\item{adjust_method}{Method fro adjusting p-values.}

\item{verbose}{Logical. Display messages}
}
\value{
A SeuratPlus object.
}
\description{
Infer a Gene Regulatory Network with \code{Pando}
}
