% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/regions.R
\name{initiate_grn}
\alias{initiate_grn}
\alias{initiate_grn.Seurat}
\title{Initiate the \code{RegulatoryNetwork} object.}
\usage{
initiate_grn(object, ...)

\method{initiate_grn}{Seurat}(
  object,
  regions = NULL,
  peak_assay = "peaks",
  rna_assay = "RNA",
  exclude_exons = TRUE
)
}
\arguments{
\item{regions}{Candidate regions to consider for binding site inference.
If \code{NULL}, all peaks regions are considered.}

\item{peak_assay}{A character vector indicating the name of the chromatin
accessibility assay in the \code{Seurat} object.}

\item{rna_assay}{A character vector indicating the name of the gene expression
assay in the \code{Seurat} object.}

\item{exclude_exons}{Logical. Whether to consider exons for binding site inference.}
}
\value{
A SeuratPlus object containing a RegulatoryNetwork object.
}
\description{
Initiate the \code{RegulatoryNetwork} object.
}
