% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{LayerData}
\alias{LayerData}
\alias{LayerData.GRNData}
\title{Get layer data from GRNData}
\usage{
LayerData(object, ...)

\method{LayerData}{GRNData}(object, ...)
}
\description{
Get layer data from GRNData
}
