% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector with ground truth values.}

\item{y_pred}{A numeric vector with predicted values.}
}
\description{
Mean Absolute Error
}
