% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_par}
\alias{map_par}
\title{Apply function over a List or Vector (in parallel)}
\usage{
map_par(x, fun, parallel = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{A vector or list to apply over.}

\item{fun}{The function to be applied to each element.}

\item{parallel}{Logical. Whether to use \code{foreach} to parallelize.}

\item{verbose}{Logical. Whether to print progress bar.
Only works in sequential mode.}
}
\value{
A list.
}
\description{
Apply function over a List or Vector (in parallel)
}
