% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percent Error}
\usage{
mape(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector with ground truth values.}

\item{y_pred}{A numeric vector with predicted values.}
}
\description{
Mean Absolute Percent Error
}
