% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Network-class}
\alias{Network-class}
\alias{Network}
\title{The Network class}
\description{
The Network object stores the inferred network itself, information about the fitting
process as well as graph representations of the network.
}
\section{Slots}{

\describe{
\item{\code{fit}}{A dataframe with goodness of fit measures.}

\item{\code{coefs}}{A dataframe with the fitted coefficients.}

\item{\code{graph}}{A graphical representation of the inferred network.}
}}

