% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{NetworkFeatures}
\alias{NetworkFeatures}
\alias{NetworkFeatures.SeuratPlus}
\title{Get network features}
\usage{
NetworkFeatures(object, ...)

\method{NetworkFeatures}{SeuratPlus}(object)
}
\description{
Get network features
}
