% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{NetworkGraph}
\alias{NetworkGraph}
\alias{NetworkGraph.SeuratPlus}
\alias{NetworkGraph.RegulatoryNetwork}
\alias{NetworkGraph.Network}
\title{Get network parameters}
\usage{
NetworkGraph(object, ...)

\method{NetworkGraph}{SeuratPlus}(object, network = DefaultNetwork(object), graph = "module_graph")

\method{NetworkGraph}{RegulatoryNetwork}(object, network = DefaultNetwork(object), graph = "module_graph")

\method{NetworkGraph}{Network}(object, graph = "module_graph")
}
\description{
Get network parameters
}
