% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{NetworkTFs}
\alias{NetworkTFs}
\alias{NetworkTFs.SeuratPlus}
\title{Get network TFs}
\usage{
NetworkTFs(object, ...)

\method{NetworkTFs}{SeuratPlus}(object)
}
\description{
Get network TFs
}
