% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Params}
\alias{Params}
\alias{Params.SeuratPlus}
\alias{Params.RegulatoryNetwork}
\title{Get GRN inference parameters}
\usage{
Params(object, ...)

\method{Params}{SeuratPlus}(object)

\method{Params}{RegulatoryNetwork}(object)
}
\description{
Get GRN inference parameters
}
