% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{plot_gof}
\alias{plot_gof}
\alias{plot_gof.SeuratPlus}
\title{Plot goodness-of-fit metrics.}
\usage{
plot_gof(object, ...)

\method{plot_gof}{SeuratPlus}(object, point_size = 0.5)
}
\value{
A ggplot2 object.
}
\description{
Plot goodness-of-fit metrics.
}
