% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{plot_network_graph}
\alias{plot_network_graph}
\alias{plot_network_graph.SeuratPlus}
\title{Plot network graph.}
\usage{
plot_network_graph(object, ...)

\method{plot_network_graph}{SeuratPlus}(
  object,
  network = DefaultNetwork(object),
  graph = "module_graph",
  layout = "umap",
  edge_width = 0.2,
  edge_color = c(`-1` = "darkgrey", `1` = "orange"),
  node_color = pals::magma(100),
  node_size = c(1, 5),
  text_size = 10,
  color_nodes = TRUE,
  label_nodes = TRUE,
  color_edges = TRUE
)
}
\arguments{
\item{object}{An object.}

\item{network}{Name of the network to use.}

\item{graph}{Name of the graph.}

\item{layout}{Layout for the graph. Can be 'umap' or any force-directed layout
implemented in \code{\link[ggraph]{ggraph}}}

\item{edge_width}{Edge width.}

\item{edge_color}{Edge color.}

\item{node_color}{Node color or color gradient.}

\item{node_size}{Node size range.}

\item{text_size}{Font size for labels.}

\item{color_nodes}{Logical, Whether to color nodes by centrality.}

\item{label_nodes}{Logical, Whether to label nodes with gene name.}

\item{color_edges}{Logical, Whether to color edges by direction.}
}
\value{
A SeuratPlus object.
}
\description{
Plot network graph.
}
