% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{plot_tf_network}
\alias{plot_tf_network}
\alias{plot_tf_network.SeuratPlus}
\title{Plot sub-network centered around one TF.}
\usage{
plot_tf_network(object, ...)

\method{plot_tf_network}{SeuratPlus}(
  object,
  tf,
  network = DefaultNetwork(object),
  graph = "module_graph",
  circular = TRUE,
  order = 3,
  edge_width = 0.2,
  edge_color = c(`-1` = "darkgrey", `1` = "orange"),
  node_color = pals::magma(100),
  node_size = c(1, 5),
  text_size = 10,
  color_nodes = TRUE,
  label_tfs = TRUE,
  color_edges = TRUE
)
}
\arguments{
\item{object}{An object.}

\item{tf}{The transcription factor to center around.}

\item{network}{Name of the network to use.}

\item{graph}{Name of the graph.}

\item{circular}{Logical. Layout tree in circular layout.}

\item{order}{Integer indicating the maximal order of the graph.}

\item{edge_width}{Edge width.}

\item{edge_color}{Edge color.}

\item{node_color}{Node color or color gradient.}

\item{node_size}{Edge size range.}

\item{text_size}{Text width.}

\item{color_nodes}{Logical, Whether to color nodes by centrality.}

\item{label_tfs}{Logical, Whether to label TFs with gene name.}

\item{color_edges}{Logical, Whether to color edges by direction.}

\item{features}{Features to use. If \code{NULL} uses all features in the graph.}
}
\value{
A SeuratPlus object.
}
\description{
Plot sub-network centered around one TF.
}
