% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{r2}
\alias{r2}
\title{\eqn{R^2} (coefficient of determination)}
\usage{
r2(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector with ground truth values.}

\item{y_pred}{A numeric vector with predicted values.}
}
\description{
\eqn{R^2} (coefficient of determination)
}
