% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Regions-class}
\alias{Regions-class}
\alias{Regions}
\title{The Regions class}
\description{
The Regions object stores the genomic regions that are considered by the model.
It stores their genomic positions, how they map to the peaks in the Seurat object
and motif matches.
}
\section{Slots}{

\describe{
\item{\code{motifs}}{A \code{Motifs} object with matches of TF motifs.}

\item{\code{ranges}}{A \code{GenomicRanges} object.}

\item{\code{peaks}}{A numeric vector with peak indices for each region.}
}}

