% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RegionStats}
\alias{RegionStats}
\alias{RegionStats.GRNData}
\title{Compute region stats from GRNData}
\usage{
RegionStats(object, ...)

\method{RegionStats}{GRNData}(object, ...)
}
\description{
Compute region stats from GRNData
}
