% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rse}
\alias{rse}
\title{Relative Squared Error}
\usage{
rse(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector with ground truth values.}

\item{y_pred}{A numeric vector with predicted values.}
}
\description{
Relative Squared Error
}
