% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RunChromVAR}
\alias{RunChromVAR}
\alias{RunChromVAR.GRNData}
\title{Run chromVAR on GRNData}
\usage{
RunChromVAR(object, ...)

\method{RunChromVAR}{GRNData}(object, ...)
}
\description{
Run chromVAR on GRNData
}
