% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{score_bagging_ridge}
\alias{score_bagging_ridge}
\title{Score a bagging ridge regression model on a test set
using a range of regression metrics.}
\usage{
score_bagging_ridge(
  formula,
  train,
  test,
  alpha = 1,
  solver = "auto",
  bagging_number = 200L,
  n_jobs = 1,
  ...
)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{train}{A \code{data.frame} containing the training data.}

\item{test}{A \code{data.frame} containing the test data.}

\item{alpha}{Positive float indicating the regularization strength.}

\item{solver}{Solver to use in the computational routines.
Options include ‘auto’, ‘svd’, ‘cholesky’, ‘lsqr’, ‘sparse_cg’, ‘sag’, ‘saga’.}

\item{bagging_number}{The number of ridge regression model in the bagging.}

\item{n_jobs}{The number of cores used to fit the model.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Score a bagging ridge regression model on a test set
using a range of regression metrics.
}
