% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{score_bayesian_ridge}
\alias{score_bayesian_ridge}
\title{Score a bagging ridge regression model on a test set
using a range of regression metrics.}
\usage{
score_bayesian_ridge(formula, train, test, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{train}{A \code{data.frame} containing the training data.}

\item{test}{A \code{data.frame} containing the test data.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Score a bagging ridge regression model on a test set
using a range of regression metrics.
}
