% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{score_cvglmnet}
\alias{score_cvglmnet}
\title{Score a cross-validated regularized generalized linear model on a test set
using a range of regression metrics.}
\usage{
score_cvglmnet(formula, train, test, alpha = 0.5, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{train}{A \code{data.frame} containing the training data.}

\item{test}{A \code{data.frame} containing the test data.}

\item{alpha}{The elasticnet mixing parameter. See \code{\link[glmnet]{glmnet}} for details.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A \code{data.frame} containing different evaluation metrics.
}
\description{
Score a cross-validated regularized generalized linear model on a test set
using a range of regression metrics.
}
