% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{score_glm}
\alias{score_glm}
\title{Score a generalized linear model on a test set using a range of regression metrics.}
\usage{
score_glm(formula, train, test, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{train}{A \code{data.frame} containing the training data.}

\item{test}{A \code{data.frame} containing the test data.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A \code{data.frame} containing different scores.
}
\description{
Score a generalized linear model on a test set using a range of regression metrics.
}
