% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{score_xgb}
\alias{score_xgb}
\title{Score a gradient boosting regression model on a test set
using a range of regression metrics.}
\usage{
score_xgb(
  formula,
  train,
  test,
  params = list(max_depth = 3, eta = 0.01, objective = "reg:squarederror"),
  nrounds = 1000,
  ...
)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{train}{A \code{data.frame} containing the training data.}

\item{test}{A \code{data.frame} containing the test data.}

\item{params}{A list with model parameters. For details, see \code{\link[xgb.train]{xgboost}}}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A \code{data.frame} containing different evaluation metrics.
}
\description{
Score a gradient boosting regression model on a test set
using a range of regression metrics.
}
