% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparse_cor}
\alias{sparse_cor}
\title{Safe correlation function which returns a sparse matrix without missing values}
\usage{
sparse_cor(
  x,
  y = NULL,
  method = "pearson",
  allow_neg = TRUE,
  remove_na = TRUE,
  remove_inf = TRUE,
  ...
)
}
\arguments{
\item{x}{Sparse matrix or character vector.}

\item{y}{Sparse matrix or character vector.}

\item{method}{Method to use for calculating the correlation coefficient.}

\item{allow_neg}{Logical. Whether to allow negative values or set them to 0.}

\item{...}{Other arguments passed to the correlation function.}
}
\value{
A correlation matrix.
}
\description{
Safe correlation function which returns a sparse matrix without missing values
}
