% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sse}
\alias{sse}
\title{Sum of Squared Errors}
\usage{
sse(y_true, y_pred)

sse(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector with ground truth values.}

\item{y_pred}{A numeric vector with predicted values.}
}
\description{
Sum of Squared Errors

Sum of Squared Errors
}
