% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{VariableFeatures.GRNData}
\alias{VariableFeatures.GRNData}
\title{Get variable features from GRNData}
\usage{
\method{VariableFeatures}{GRNData}(object, ...)
}
\description{
Get variable features from GRNData
}
