% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_assay}
\alias{aggregate_assay}
\title{Aggregate Seurat assay over groups}
\usage{
aggregate_assay(
  object,
  group_name,
  fun = "mean",
  assay = "RNA",
  layer = "data",
  slot = NULL
)
}
\arguments{
\item{group_name}{A character vector indicating the metadata column to aggregate over.}

\item{fun}{The summary function to be applied to each group.}

\item{assay}{The assay to summarize.}

\item{layer}{The layer to summarize (Seurat v5).}

\item{slot}{Deprecated. Use \code{layer} instead.}
}
\value{
A Seurat object.
}
\description{
Aggregate Seurat assay over groups
}
