% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{coef.SeuratPlus}
\alias{coef.SeuratPlus}
\alias{coef.RegulatoryNetwork}
\alias{coef.Network}
\title{Get fitted coefficients}
\usage{
\method{coef}{SeuratPlus}(object, network = "glm_network")

\method{coef}{RegulatoryNetwork}(object, network = "glm_network")

\method{coef}{Network}(object)
}
\description{
Get fitted coefficients
}
