% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/grn.R
\name{find_modules}
\alias{find_modules}
\alias{find_modules.Network}
\alias{find_modules.SeuratPlus}
\title{Find TF modules in regulatory network}
\usage{
find_modules(object, ...)

\method{find_modules}{Network}(
  object,
  p_thresh = 0.05,
  rsq_thresh = 0.1,
  nvar_thresh = 10,
  min_genes_per_module = 5,
  xgb_method = c("tf", "target"),
  xgb_top = 50,
  verbose = TRUE
)

\method{find_modules}{SeuratPlus}(
  object,
  network = DefaultNetwork(object),
  p_thresh = 0.05,
  rsq_thresh = 0.1,
  nvar_thresh = 10,
  min_genes_per_module = 5
)
}
\arguments{
\item{object}{An object.}

\item{p_thresh}{Float indicating the significance threshold on the adjusted p-value.}

\item{rsq_thresh}{Float indicating the \eqn{R^2} threshold on the adjusted p-value.}

\item{nvar_thresh}{Integer indicating the minimum number of variables in the model.}

\item{min_genes_per_module}{Integer indicating the minimum number of genes in a module.}

\item{xgb_method}{Method to get modules from xgb models
\itemize{
\item \code{'tf'} - Choose top targets for each TF.
\item \code{'target'} - Choose top TFs for each target gene.
}}

\item{xgb_top}{Interger indicating how many top targets/TFs to return.}

\item{verbose}{Print messages.}

\item{network}{Name of the network to use.}
}
\value{
A Network object.

A SeuratPlus object.
}
\description{
Find TF modules in regulatory network
}
