% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_bayesian_ridge}
\alias{fit_bayesian_ridge}
\title{Fit a bayesian ridge regression model as implemented in scikit-learn (python)}
\usage{
fit_bayesian_ridge(formula, data, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Fit a bayesian ridge regression model as implemented in scikit-learn (python)
}
