% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{get_network_graph}
\alias{get_network_graph}
\alias{get_network_graph.SeuratPlus}
\title{Compute network graph embedding using UMAP.}
\usage{
get_network_graph(object, ...)

\method{get_network_graph}{SeuratPlus}(
  object,
  network = "glm_network",
  graph_name = "module_graph",
  rna_assay = "RNA",
  rna_slot = "data",
  use_coexpression = TRUE,
  features = NULL,
  random_seed = 111
)
}
\value{
A SeuratPlus object.
}
\description{
Compute network graph embedding using UMAP.
}
