% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{get_network_graph}
\alias{get_network_graph}
\alias{get_network_graph.SeuratPlus}
\title{Compute network graph embedding using UMAP.}
\usage{
get_network_graph(object, ...)

\method{get_network_graph}{SeuratPlus}(
  object,
  network = DefaultNetwork(object),
  graph_name = "module_graph",
  rna_assay = "RNA",
  rna_slot = "data",
  umap_method = c("weighted", "corr", "coef", "none"),
  features = NULL,
  random_seed = 111,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments for \code{\link[umap]{uwot}}.}

\item{network}{Name of the network to use.}

\item{graph_name}{Name of the graph.}

\item{rna_assay}{Name of the RNA assay.}

\item{rna_slot}{Name of the RNA slot to use.}

\item{umap_method}{Method to compute edge weights for UMAP:
\itemize{
\item \code{'weighted'} - Correlation weighted by GRN coefficient.
\item \code{'corr'} - Only correlation.
\item \code{'coef'} - Only GRN coefficient.
\item \code{'none'} - Don't compute UMAP and create the graph directly
from modules.
}}

\item{features}{Features to use to create the graph. If \code{NULL}
uses all features in the network.}

\item{random_seed}{Random seed for UMAP computation}

\item{verbose}{Print messages.}
}
\value{
A SeuratPlus object.
}
\description{
Compute network graph embedding using UMAP.
}
