% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{GetNetwork}
\alias{GetNetwork}
\alias{GetNetwork.GRNData}
\alias{GetNetwork.RegulatoryNetwork}
\title{Get network}
\usage{
GetNetwork(object, ...)

\method{GetNetwork}{GRNData}(object, network = DefaultNetwork(object))

\method{GetNetwork}{RegulatoryNetwork}(object, network = DefaultNetwork(object))
}
\description{
Get network

Get network
}
