% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{gof}
\alias{gof}
\alias{gof.SeuratPlus}
\alias{gof.RegulatoryNetwork}
\title{Get goodness-of-fit info}
\usage{
gof(object, ...)

\method{gof}{SeuratPlus}(object)

\method{gof}{RegulatoryNetwork}(object)
}
\description{
Get goodness-of-fit info
}
