% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{gof}
\alias{gof}
\alias{gof.SeuratPlus}
\alias{gof.RegulatoryNetwork}
\alias{gof.Network}
\title{Get goodness-of-fit info}
\usage{
gof(object, ...)

\method{gof}{SeuratPlus}(object, network = DefaultNetwork(object))

\method{gof}{RegulatoryNetwork}(object, network = DefaultNetwork(object))

\method{gof}{Network}(object, network = DefaultNetwork(object))
}
\description{
Get goodness-of-fit info
}
