% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/grn.R
\name{infer_grn}
\alias{infer_grn}
\alias{infer_grn.SeuratPlus}
\title{Infer a Gene Regulatory Network with \code{Pando}}
\usage{
infer_grn(object, ...)

\method{infer_grn}{SeuratPlus}(
  object,
  genes = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  parallel = FALSE,
  tf_cor = 0.1,
  peak_cor = 0,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("glm", "glmnet", "cv.glmnet", "brms", "xgb", "bagging_ridge",
    "bayesian_ridge"),
  alpha = 0.5,
  family = "gaussian",
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{...}{Other parameters for the model fitting function.}

\item{genes}{A character vector with the target genes to consider for GRN inference.
Takes all VariableFeatures in the object per default.}

\item{peak_to_gene_method}{Character specifying the method to
link peak overlapping motif regions to nearby genes. One of 'Signac' or 'GREAT'.}

\item{upstream}{Integer defining the distance upstream of the gene to consider as potential regulatory region.}

\item{downstream}{Integer defining the distance downstream of the gene to consider as potential regulatory region.}

\item{extend}{Integer defining the distance from the upstream and downstream of the basal regulatory region.
Only used of \code{peak_to_gene_method = 'GREAT'}.}

\item{only_tss}{Logical. Measure distance from the TSS (\code{TRUE}) or from the entire gene body (\code{FALSE}).}

\item{parallel}{Logical. Whether to parallelize the computation with \code{\link[foreach]{foreach}}.}

\item{tf_cor}{Threshold for TF - target gene correlation.}

\item{peak_cor}{Threshold for binding peak - target gene correlation.}

\item{method}{A character string indicating the method to fit the model.
\itemize{
\item \code{'glm'} - Generalized Liner Model with \code{\link[glm]{stats}}.
\item \code{'glmnet'}, \code{'cv.glmnet'} - Regularized Generalized Liner Model with \code{\link[glmnet]{glmnet}}.
\item \code{'brms'} - Bayesian Regression Models using \code{\link[brms-package]{brms}}.
\item \code{'xgb'} - Gradient Boosting Regression using \code{\link[xgboost]{xgboost}}.
\item \code{'bagging_ridge'} - Bagging Ridge Regression using scikit-learn via \link[reticulate]{reticulate}.
\item \code{'bayesian_ridge'} - Bayesian Ridge Regression using scikit-learn via \link[reticulate]{reticulate}.
}}

\item{alpha}{The elasticnet mixing parameter. See \code{\link[glmnet]{glmnet}} for details.}

\item{family}{A description of the error distribution and link function to be used in the model.
See \code{\link[family]{stats}} for mode details.}

\item{interaction_term}{The interaction term to use in the model between TF and binding site.
\itemize{
\item \code{'+'} for additive interaction.
\item \code{':'} for 'multiplicative' interaction.
\item \code{'*'} for crossing interaction, i.e. additive AND 'multiplicative'.
For more info, see \code{\link[formula]{stats}}
}}

\item{adjust_method}{Method for adjusting p-values.}

\item{scale}{Logical. Whether to z-transform the expression and accessibility matrices.}

\item{verbose}{Logical. Display messages. Set verbose to '2' to print errors for all model fits.}
}
\value{
A SeuratPlus object.
}
\description{
Infer a Gene Regulatory Network with \code{Pando}
}
