% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{NetworkModules}
\alias{NetworkModules}
\alias{NetworkModules.SeuratPlus}
\alias{NetworkModules.RegulatoryNetwork}
\alias{NetworkModules.Network}
\title{Get TF modules}
\usage{
NetworkModules(object, ...)

\method{NetworkModules}{SeuratPlus}(object, network = DefaultNetwork(object))

\method{NetworkModules}{RegulatoryNetwork}(object, network = DefaultNetwork(object))

\method{NetworkModules}{Network}(object)
}
\description{
Get TF modules
}
