% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{NetworkParams}
\alias{NetworkParams}
\alias{NetworkParams.SeuratPlus}
\alias{NetworkParams.RegulatoryNetwork}
\alias{NetworkParams.Network}
\title{Get network parameters}
\usage{
NetworkParams(object, ...)

\method{NetworkParams}{SeuratPlus}(object, network = "glm_network")

\method{NetworkParams}{RegulatoryNetwork}(object, network = "glm_network")

\method{NetworkParams}{Network}(object)
}
\description{
Get network parameters
}
