% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{NetworkParams}
\alias{NetworkParams}
\alias{NetworkParams.SeuratPlus}
\title{Get network parameters}
\usage{
NetworkParams(object, ...)

\method{NetworkParams}{SeuratPlus}(object)
}
\description{
Get network parameters
}
