% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{NetworkRegions}
\alias{NetworkRegions}
\alias{NetworkRegions.SeuratPlus}
\alias{NetworkRegions.RegulatoryNetwork}
\title{Get network regions}
\usage{
NetworkRegions(object, ...)

\method{NetworkRegions}{SeuratPlus}(object)

\method{NetworkRegions}{RegulatoryNetwork}(object)
}
\description{
Get network regions
}
