% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{NetworkTFs}
\alias{NetworkTFs}
\alias{NetworkTFs.GRNData}
\alias{NetworkTFs.RegulatoryNetwork}
\title{Get network TFs}
\usage{
NetworkTFs(object, ...)

\method{NetworkTFs}{GRNData}(object)

\method{NetworkTFs}{RegulatoryNetwork}(object)
}
\description{
Get network TFs
}
