% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{plot_tf_network}
\alias{plot_tf_network}
\alias{plot_tf_network.SeuratPlus}
\title{Plot sub-network centered around one TF.}
\usage{
plot_tf_network(object, ...)

\method{plot_tf_network}{SeuratPlus}(
  object,
  tf,
  network = DefaultNetwork(object),
  graph = "module_graph",
  circular = TRUE,
  edge_width = 0.2,
  edge_color = c(`-1` = "darkgrey", `1` = "orange"),
  node_size = 3,
  text_size = 10,
  label_nodes = c("tfs", "all", "none"),
  color_edges = TRUE
)
}
\arguments{
\item{object}{An object.}

\item{tf}{The transcription factor to center around.}

\item{network}{Name of the network to use.}

\item{graph}{Name of the graph.}

\item{circular}{Logical. Layout tree in circular layout.}

\item{edge_width}{Edge width.}

\item{edge_color}{Edge color.}

\item{node_size}{Node size.}

\item{text_size}{Font size for labels.}

\item{label_nodes}{String, indicating what to label.
\itemize{
\item \code{'tfs'} - Label all TFs.
\item \code{'all'} - Label all genes.
\item \code{'none'} - Label nothing (except the root TF).
}}

\item{color_edges}{Logical, whether to color edges by direction.}
}
\value{
A SeuratPlus object.
}
\description{
Plot sub-network centered around one TF.
}
