% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{VariableFeatures}
\alias{VariableFeatures}
\alias{VariableFeatures.GRNData}
\title{Get variable features from GRNData}
\usage{
VariableFeatures(object, ...)

\method{VariableFeatures}{GRNData}(object, ...)
}
\description{
Get variable features from GRNData
}
